created setup.py
created the setup.py script witch is used by pypi to create the reposetory
