+++ make.bat
@@ -1,0 +1,6 @@
+@echo off
+call cls
+echo.running setup
+echo.
+
+python36 setup.py sdist
